requirements.txt
SpeechRecognition
pyautogui
keyboard
PyAudio
pynput==1.7.6
requests==2.31.0
psutil==5.9.8
infi.systray==0.1.12 